+++ built_value_generator/lib/src/strings.dart
@@ -1,0 +1,2 @@
+/// Escapes dollar signs in a string with backslashes.
+String escapeString(String string) => string.replaceAll(r'$', r'\$');